+++ nodeconductor/structure/serializers.py
@@ -197,7 +197,8 @@ class NestedServiceProjectLinkSerializer(serializers.Serializer):
         return total
 
 
-class ProjectSerializer(PermissionFieldFilteringMixin,
+class ProjectSerializer(core_serializers.RestrictedSerializerMixin,
+                        PermissionFieldFilteringMixin,
                         core_serializers.AugmentedSerializerMixin,
                         serializers.HyperlinkedModelSerializer):
     project_groups = NestedProjectGroupSerializer(
@@ -316,7 +317,8 @@ class CustomerImageSerializer(serializers.ModelSerializer):
         fields = ['image']
 
 
-class CustomerSerializer(core_serializers.AugmentedSerializerMixin,
+class CustomerSerializer(core_serializers.RestrictedSerializerMixin,
+                         core_serializers.AugmentedSerializerMixin,
                          serializers.HyperlinkedModelSerializer,):
     projects = PermissionProjectSerializer(many=True, read_only=True)
     project_groups = PermissionProjectGroupSerializer(many=True, read_only=True)
@@ -1343,7 +1345,7 @@ class BaseResourceImportSerializer(PermissionFieldFilteringMixin,
                 {'backend_id': "This resource is already linked to NodeConductor"})
 
         spl_class = SupportedServices.get_related_models(self.Meta.model)['service_project_link']
-        spl = spl_class.objects.get(service=self.context['service'], project=validated_data.pop('project'))
+        spl = spl_class.objects.get(service=self.context['service'], project=validated_data['project'])
         validated_data['service_project_link'] = spl
 
         return validated_data